+++ example/lib/main.dart
@@ -84,10 +84,14 @@ class SmartDialogPage extends StatelessWidget {
     var attachDialog = (BuildContext context) {
       SmartDialog.showAttach(
         targetContext: context,
-        alignment: Alignment.topCenter,
-        useSystem: true,
+        alignment: Alignment.bottomCenter,
+        animationType: SmartAnimationType.scale,
+        scalePoint: Offset(100, 0),
         builder: (_) {
-          return Container(width: 100, height: 100, color: Colors.red);
+          return Column(mainAxisSize: MainAxisSize.min, children: [
+            Container(width: 100, height: 50, color: Colors.red),
+            Container(width: 100, height: 50, color: Colors.blue),
+          ]);
         },
       );
     };