+++ CHANGELOG.md
@@ -5,6 +5,20 @@
 - Fix [#56](https://github.com/fluttercandies/flutter_smart_dialog/issues/56)
 - Fix [#60](https://github.com/fluttercandies/flutter_smart_dialog/issues/60)
 - Add SmartDialogController
+- adjust AnimationType
+  - fade: FadeTransition for all positions
+  - scale: All positions are ScaleTransition
+  - centerFade_otherSlide: The center position is the FadeTransition, and the other positions are the SlideTransition
+  - centerScale_otherSlide: The center position is the ScaleTransition, and the other positions are the SlideTransition
+- add scalePoint (showAttach)
+
+# [4.0.5]
+- adjust AnimationType
+    - fade: FadeTransition for all positions
+    - scale: All positions are ScaleTransition
+    - centerFade_otherSlide: The center position is the FadeTransition, and the other positions are the SlideTransition
+    - centerScale_otherSlide: The center position is the ScaleTransition, and the other positions are the SlideTransition
+- add scalePoint (showAttach)
 
 # [4.0.3]